+++ server/ngwmn/__init__.py
@@ -3,7 +3,7 @@ Initialize the NGWMN UI application
 """
 import json
 
-from flask import Flask
+from flask import Flask, render_template
 
 
 __version__ = '0.3.0dev'
@@ -23,6 +23,19 @@ manifest_path = app.config.get('ASSET_MANIFEST_PATH')
 if manifest_path:
     with open(manifest_path, 'r') as f:
         app.config['ASSET_MANIFEST'] = json.loads(f.read())
+
+
+@app.errorhandler(404)
+@app.errorhandler(500)
+def handle_service_exception(error):
+    return render_template(
+        'error.html',
+        error={
+            'message': error.description,
+            'status_code': error.code,
+            'status_code_str': error.name
+        }
+    ), error.code
 
 
 from . import views # pylint: disable=C0413