+++ server/ngwmn/services/__init__.py
@@ -1,4 +1,6 @@
-from flask import jsonify
+from http.client import responses
+
+from flask import render_template
 
 from ngwmn import app
 
@@ -22,7 +24,8 @@ class ServiceException(Exception):
         return {
             **self.payload,
             'message': self.message,
-            'status_code': self.status_code
+            'status_code': self.status_code,
+            'status_code_str': responses[self.status_code]
         }
 
 
@@ -31,6 +34,7 @@ def handle_service_exception(error):
     """
     Capture raised ServiceExceptions and return the appropriate status code.
     """
-    response = jsonify(error.to_dict())
-    response.status_code = error.status_code
-    return response
+    return render_template(
+        'error.html',
+        error=error.to_dict()
+    ), error.status_code